refactor(Player, BossMonster): add isVictory method
Player, BossMonster 클래스에 승리여부 확인 메소드 추가
